move variable
